+++ commons/src/main/java/com/karuslabs/commons/command/synchronization/SynchronizationEvent.java
@@ -32,7 +32,7 @@ import org.bukkit.event.player.PlayerCommandSendEvent;
 
 public class SynchronizationEvent extends PlayerCommandSendEvent {
     
-    private static final HandlerList handlers = new HandlerList();
+    private static final HandlerList HANDLERS = new HandlerList();
     
     
     public SynchronizationEvent(Player player, Collection<String> commands) {
@@ -42,11 +42,11 @@ public class SynchronizationEvent extends PlayerCommandSendEvent {
     
     @Override
     public HandlerList getHandlers() {
-        return handlers;
+        return HANDLERS;
     }
 
     public static HandlerList getHandlerList() {
-        return handlers;
+        return HANDLERS;
     }
     
 }