+++ commons/src/test/java/com/karuslabs/commons/command/synchronization/SynchronizerTest.java
@@ -109,7 +109,7 @@ class SynchronizerTest {
     
     @Test
     void synchronize_player_commands() {
-        synchronizer.tree = mock(TreeWalker.class);
+        synchronizer.walker = mock(TreeWalker.class);
         
         EntityPlayer entity = mock(EntityPlayer.class);
         entity.playerConnection = mock(PlayerConnection.class);
@@ -118,7 +118,7 @@ class SynchronizerTest {
         
         synchronizer.synchronize(player, List.of());
         
-        verify(synchronizer.tree).add(any(), any(), any(), any());
+        verify(synchronizer.walker).add(any(), any(), any(), any());
         verify(entity.playerConnection).sendPacket(any(PacketPlayOutCommands.class));
     }